+++ pokemon gone mad/Assets/Scenes/GameplayScene(For Real).unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e95947644c2b0e74084ae4a1671ad3e4
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
