+++ pokemon gone mad/Assets/Scripts/PlayerScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2781e6f93ed06f44d8ae9fe75e8d361d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
