+++ pokemon gone mad/Assets/Scripts/PlayerScripts/HeatManager.cs.meta
