+++ pokemon gone mad/Assets/Scripts/PlayerScripts/Weapon.cs
