+++ pokemon gone mad/Assets/Scripts/RoomManagement.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 015b73a87e95c5243876d0b4d114716e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
